+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added EMI support
 * Fixed Sneak + Right Click in the Mycelial Reactor not showing which generators aren't working
 * Fixed Myceliar Reactor Having the render enabled always
+* Added a button to disable particle spawning in machines with a working area
 
 # Version 3.6.11